standardize on 'int' for message sizes
that's what the sources already assumed anyway. size_t is total
overkill, as No Email Ever (TM) will exceed 2GiB.

this also fixes a harmless format string warning in 32 bit builds.

+++ src/driver.h
@@ -39,7 +39,7 @@ typedef struct store_conf {
 	const char *flat_delim;
 	const char *map_inbox;
 	const char *trash;
-	uint max_size; /* off_t is overkill */
+	int max_size; /* off_t is overkill */
 	char trash_remote_new, trash_only_new;
 } store_conf_t;
 
@@ -64,7 +64,7 @@ typedef struct message {
 	struct message *next;
 	struct sync_rec *srec;
 	/* string_list_t *keywords; */
-	size_t size; /* zero implies "not fetched" */
+	int size; /* zero implies "not fetched" */
 	int uid;
 	uchar flags, status;
 	char tuid[TUIDL];
